# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce API Documentation",
        "description": "Documentación interactiva para la API del sistema de E-commerce, desarrollada para la prueba técnica de Zenova Digital.",
        "contact": {
            "email": "leidy.palomino@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor Principal de la API"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar y filtrar productos",
                "description": "Obtiene una lista paginada de productos con opciones avanzadas de filtrado y ordenamiento.",
                "operationId": "ab015cc2055ac75e6d23c628675c26e7",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda full-text",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "ID de la categoría para filtrar",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Precio mínimo",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Precio máximo",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "draft"
                            ]
                        }
                    },
                    {
                        "name": "featured",
                        "in": "query",
                        "description": "Filtrar por productos destacados",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Campo para ordenar",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "price",
                                "name",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Dirección de ordenamiento",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "Incluir relaciones (separadas por coma, ej: categories,images)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de productos paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Product"
                                            }
                                        },
                                        "path": {
                                            "type": "string"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "next_cursor": {
                                            "type": "string"
                                        },
                                        "next_page_url": {
                                            "type": "string"
                                        },
                                        "prev_cursor": {
                                            "type": "string"
                                        },
                                        "prev_page_url": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un nuevo producto en la base de datos (Requiere rol de 'admin').",
                "operationId": "7ff81b1681ff3180e5b73a2e15f2c67b",
                "requestBody": {
                    "description": "Datos del nuevo producto",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado (no es admin)"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener un producto por su ID",
                "description": "Devuelve los detalles completos de un solo producto.",
                "operationId": "47affec446a4e21ba3cbf0b6b14abb44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar un producto existente",
                "description": "Actualiza los datos de un producto (Requiere rol de 'admin').",
                "operationId": "b0605b594dede1c8d6d6412b452efd96",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Eliminar un producto",
                "description": "Elimina un producto de la base de datos (Requiere rol de 'admin').",
                "operationId": "33173a96c7ab5ed4ba9ceaffb4c6003d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado exitosamente"
                    },
                    "403": {
                        "description": "Acceso denegado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}/images": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Subir una imagen para un producto",
                "description": "Sube un archivo de imagen y lo asocia a un producto (Requiere rol de 'admin').",
                "operationId": "0baccabdf5f6572064b668d77d119478",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "El archivo de imagen a subir.",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "alt_text": {
                                        "description": "Texto alternativo para la imagen.",
                                        "type": "string"
                                    },
                                    "is_primary": {
                                        "description": "Marcar como imagen principal.",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Imagen subida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductImage"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category Schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "name": {
                        "type": "string",
                        "example": "Laptops"
                    },
                    "slug": {
                        "type": "string",
                        "example": "laptops"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product Schema",
                "required": [
                    "name",
                    "sku",
                    "price",
                    "stock"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "sku": {
                        "description": "Stock Keeping Unit único",
                        "type": "string",
                        "example": "LP-PRO-01"
                    },
                    "name": {
                        "description": "Nombre del producto",
                        "type": "string",
                        "example": "Laptop Pro M3"
                    },
                    "slug": {
                        "type": "string",
                        "readOnly": true,
                        "example": "laptop-pro-m3"
                    },
                    "description": {
                        "type": "string",
                        "example": "Una laptop muy potente para profesionales."
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1599.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 50
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "draft"
                        ],
                        "example": "active"
                    },
                    "featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductImage"
                        }
                    }
                },
                "type": "object"
            },
            "ProductImage": {
                "title": "Product Image Schema",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://example.com/images/product.jpg"
                    },
                    "alt_text": {
                        "type": "string",
                        "example": "Imagen de una laptop"
                    },
                    "is_primary": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación con Token JWT (Bearer Token)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}